skeletons monsters

+++ simple3d/Scene/ISceneRenderer.cs
@@ -5,6 +5,6 @@ namespace simple3d.Scene
 {
     public interface ISceneRenderer: IDisposable
     {
-        void Render(IScreen screen, Level level, double elapsedMilliseconds);
+        void Render(IScreen screen, Level level, float elapsedMilliseconds);
     }
 }
